pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dmpo"
version = "1.0.0"
description = "Dispersive MeanFlow Policy Optimization: One-step flow matching policy with online RL fine-tuning."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "av==12.3.0",
    "einops==0.8.0",
    "gdown==5.2.0",
    "gym==0.22.0",
    "hydra-core==1.3.2",
    "imageio==2.35.1",
    "matplotlib==3.7.5",
    "omegaconf==2.3.0",
    "pretty_errors==1.2.25",
    "torch>=2.4.0",
    "tqdm==4.66.5",
    "wandb==0.17.7",
    "pandas==2.0.3",
    "pypdf>=3.9.0",
    "seaborn==0.13.2"
]

[project.optional-dependencies]
gym = [
    "cython<3",
    "d4rl",
    "patchelf",
]
kitchen = [
    "cython<3",
    "d4rl",
    "dm_control==1.0.16",   
    "mujoco==3.1.6",
    "patchelf",
]
robomimic = [
    "cython<3",
    "d4rl",
    "patchelf",
    "robomimic @ git+https://github.com/ARISE-Initiative/robomimic.git",
    "robosuite @ git+https://github.com/ARISE-Initiative/robosuite.git@v1.4.1",
]
d3il = [
    "cython<3",
    "d4rl",
    "gin-config",
    "patchelf",
    "pin",
]
furniture = [
    "cython<3",
    "d4rl",
    "ipdb==0.13.13",
    "patchelf",
]
exact = [   # exact likelihood
    "torchdiffeq",
]
all = [
    "cython<3",
    "d4rl",
    "gin-config",
    "ipdb==0.13.13",
    "patchelf",
    "pin",
    "robomimic @ git+https://github.com/ARISE-Initiative/robomimic.git",
    "robosuite @ git+https://github.com/ARISE-Initiative/robosuite.git@v1.4.1",
]

[project.urls]
Homepage = "https://github.com/xxx/DMPO"
Documentation = "https://github.com/xxx/DMPO/tree/main/docs"

[tool.setuptools.packages.find]
exclude = []

[tool.wheel]
exclude = []
